Update unit tests to match improved workflow
There was a minor error in the original processGroups function which
meant that GroupWorkspaces was being called twice to group the outputs
when the input is a group workspace (including when the input has been
sliced). The refactoring in this PR has removed the duplicate step so
this commit removes the additional GroupWorkspaces expectation on the
output history.

Re #23471