Store ILL run date in start_time
To be coherent in all of ILL loaders, the "run_start" parameter is
changed to "start_time".
Because it is needed before all the parameters are loaded, the start time
needs to be loaded separately. Since this would conflict with its later
add, a change is made so that it is overwritten in the parameters
loading phase.
Then, tests are modified accordingly.