Enable updating of IFunctionModel attributes from IFunctionView
This commit the model to be updated when individual attributes are changed in the view. Previously, any attribute being changed in the view would cause the full function be replaced and reconstructed. This is sometimes necessary when one of the attributes modifies the full function, e.g the order of polynomial - however most the time it is unnecessary. To avoid this a check has been included, which ensures the function is only reconstructed when it needs to be.