Enable updating of IFunctionView from IFunctionModel
This commit ensures the correct value of the attribute (as taken from the model) is shown in the IFunctionView. This issue is prevalent for Q dependent functions, with the IFunctionView typically showing the incorrect value. In order to remedy this issue, methods to set attributes in the view have been added.