Abins: tweak progress reporting
- Fix a bug whereby progress was not reported if frequencies are
  batched into chunks

- Include q-point in progress log

Given a test case with many q-points, the progress bar only seems to
increment every few q-points and therefore gives the impression that
it is not incrementing atom number. Presumably there is some kind of
rate-limiting in the logger? It might be reassuring to tone this down
so that we can see all atoms are included. Alternatively, the progress
reports could be reduced to "per-qpt" if there are many q-points.

In the near future we will need to think about how progress reporting
interacts with parallelism.