Fix hard-coded values in test which change due to changes to fwhm
The fwhm of the B2Bexp determines the range over which the data in the FitPowderDiffPeaks is fitted. This means the chisq values of the fit were different (actually improved) in addition the changes meant that the intial parameters for the function were guessed more accurately and an additional peak was fitted - so have changed the index of the reuslts table querried so as to check the same peak.
Have also made the test for the difference = data in region not fitted more clear and concise (had to change the index of last point which was inside the fitting region with new fwhm).