Fix tests with new values to account for different starting guess
Fit is now started from slightly different intial parameters. Fro the doc test the new and old fit values are consistent within the uncertainty on the fit. For the other test the windows build gave the same answers but have increased the tolerance to pass OSX and RHEl builds which now give slightly different values.