Include search results when save/load batch
Search results were not originally included in the encoder/decoder
because they can be reproduced from the search criteria. However, we now
need to include them to save the added comments

Note that we save the cached search criteria rather than what is displayed
on the view. This is because these criteria relate to the search results.
This may be a little confusing but is less so than saving search
criteria that are different to the saved search results, which can lead
to unexpected behaviour on re-loading the batch.

Re #2918

Include instrument name in encoder