Update tests and move all logic about warnings to main window
- Update tests for the new search logic

- The batch and runs presenter had some knowledge about checking whether
warnings are enabled in the options dialog. We only really need to know
about this in the main window, so generalise the code in the other
presenters so they do not need to access it. The reorganised checks mean that checks on the warn-changes flag are not
done if we've already established that the batch is saved, so the tests
are updated to reflect this

- Also remove some duplication from the tests and tidy up/remove some
misleading calls e.g. some ON_CALL statements were unnecessary

Re #29148