Remove A/B error check for valid peak in calibration
A/B parameters are fit with zero error now as they are fixed (set in parameter file).
Have corrected a bug in allowed tolerance on width error for acceptable peak - this broke a lot of tests so have changed the simulated test data for more realistic parameters for which the parameters S,A,B are less correlated.