Fix a crash loading files without pulse_time_zero
This is a regression caused by PR #29105. The implementation was changed and
no longer checks that values exist before attempting to use them. This
commit simply adds a check and throws an error if they don't exist.

This is actually an improvement over the behaviour in v4.0 - the load
used to throw an error from somewhere else in the code and
gave a confusing error about getData failing.