Return shared_pointer in Python WorkspaceGroup.getItem()
Resolves an issue where a weak pointer is returned from
WorkspaceGroup.getItem(n), this causes the workspace to fall out of
scope with the WS Group, even if packed into another list.

Originally the behaviour is to match the ADS using weak pointers, so
that lifetime is tied the ADS object. However, in this case getting
items from WorkspaceGroups using .getItems will now live as long as
there is a Python reference if the object was not in the ADS

When the object exists in the ADS we decay to a weak pointer instead,
which ties the lifetime to the data service instance. This prevents a
memory leak where a user will reference the file in a script from the
GUI, then delete the workspace from the ADS. As the script would still
hold a ref Mantid would appear to leak memory, which would subvert user
expectations