Handle unsaved search results separately to batch content
Previously we marked the whole batch as unsaved if there are unsaved
annotations in the search results. However, users may reset the search
results when performing a new search and in this case it does not make
sense to continue marking the batch as unsaved (and popping up spurious
warnings when they have already cleared their unsaved annotations).

This commit therefore adds a separate flag for unsaved search results,
handled by the search model.

Re #29148