Reset search results when search criteria are changed
Previously, resetting the search results was being done when the search
was executed, using cached criteria to check whether there had been any
changes and a reset was required. This commit changes it so that we detect when the search
criteria are changed and reset the results straight away. This avoids
having a mismatch between the displayed results and criteria.

This is especially important if the batch is saved with that mismatched
criteria, because on reloading the batch the mismatch still exists, but
the previously cached results are not available, so we cannot identify
that the search results need to be reset, leading to misleading results.

Re #29148