Creating mantid image and canvas class
This commit creates a Mantid image and canvas class. This allows us to override behaviours on the canvas class, e.g the pen colour used to draw selection rectangles (zoom option on toolbar).

The image class is created so we can create a unified method to change the pen colour based on how dark, or light, the image is. This will ensure that the matplotlib tools are visible.