Move search criteria into a struct
Tidy the search criteria in the QtCatalogSearcher into a single struct. The same struct is used in the RunsPresenter to cache the search criteria used for the last autoreduction process. This is so that we can check if the autoreduction has changed so we know whether to reset the main table when a new autoreduction is started.

Re #29148