Cache last used autoreduction search criteria
This allows us to check with the user about discarding changes before we
clear the main table. This needs to be done separately from the check to
clear the search results, which may be done at a different time. We
could perhaps clear the table when we do a normal search, but this may
be excessive.

Re #29148