Rearrange plot options window
Group axis related options in a tab view
Move label to the top of axis options
Move canvas color to be with whole axes options

Use QTabBar instead of QTabWidget, because otherwise the axis options would need to be duplicated onto each tab page. Needs to be added in the view.py as QTabBar not available in QT Designer.