Use cached search criteria for save
To avoid the problem of the search criteria that are saved being
different to the ones that were used to produce the current search results, save
the cached criteria rather than the values currently on the view. This
is slightly confusing because on Load you may get different values to
what were displayed when you Saved. However, I think this is less
confusing than having criteria that perform a different search to the
results that are loaded.

I had previously tried to avoid having to do this by clearing the search
results when the criteria are actually changed, but I think this is
annoying from a users perspective and it is difficult to get the various
combinations of signals to work nicely together (e.g. when text is
edited, enter pressed, search button pressed, autoprocess button
pressed, and we can have multiple of these signals emitted at the same
time).

Re #29148