Abins SDataByAngle container
Container for SData at multiple angles, for ease of summation /
analysis.

For now this class and SData still use the ['s'] level of nested
dicts, but this is earmarked for removal.

The class instantiation and methods are tested; I haven't checked
coverage but I think most of it is there.