Abins: remove loop over incident energy
Currently, Abins calculations for direct instruments will loop over
several incident energies and sum the results together.

It is not clear that this is especially useful, and has not been
directly requested by users. In an experimental scenario we would
expect several spectra to be collected at different incident energies,
and compared with simulations at the same energies.

This removes one level of looping/nesting from Abins, and may allow us
to further remove some of the redundant code separating 1-D and 2-D
runs.