Make MantidPlot opt in for new build folders
Makes MantidPlot opt-in for new CMake generations. At the point of
writing the CI jobs have been building with MantidPlot disabled fine for
a while.

Make it so that any new CMake build directories have it off by default,
so people can opt-in if they really need it. Note this does not impact
existing CMakeCaches, so someone with Plot on will continue to have it
on.