Removing linking to Python libraries directly
Conda statically links the Python executable to
the runtime library. If we also link dynamically to
the Python runtime then there is a symbol
clash and the dynamic symbols overwrite the
static ones leading to errors like:
SystemError: initialization of _kernel raised
unreported exception or a segfault in
PyUnicode_InternInPlace. https://github.com/shogun-toolbox/shogun/issues/4068