Fix an error in calculating the projected lambda range
The projection calculation assumes a bin centre is given but for the
overall min/max lambda we are using the values the user wants to crop
to, which are bin edges. This commit fixes that so the final output
range is a projection of the correct input crop range.

There was also a bug in calculating the bin width for this calculation.
It had an off-by-one error in the number of bins used to calculate the
bin width (one too many bins, hence the bin width was slightly smaller
than it should have been). This commit also fixes that.

I think the only effect of this bug was that the output range was being
cropped to a slightly larger range that it should have been, which
should not affect the validity of the results, except perhaps the bins
at the edges may only be partially filled. It will also affect the
binning of the output workspace, which is based on the same bin width as
the input workspace, so the number of bins in the output will be
different as will the exact position of the bin edges.

Re #23012