Fix ISIS SANS Fitting not doing anything
Fixes the SANS GUI models not actually changing the underlying state.

The original SANS design used a dict (which mapped to AllStates) to
manipulate the settings, which were passed to a parser. Now we have a
common shared AllStates, so in effect the GUI models are a wrapper layer
on the underlaying States Internal objects.

Because of this legacy we still were creating deep copies, of which the
Adjustment Model ended up discarded, causing it to appear like nothing
changed.
Since we are only using the models as wrappers around all states we no
longer need to create copies anymore, fixing the problem.