Prevent falsey types being reset back to default in ISIS SANS
Falsey types, such as False or 0.0 were being picked up in multiple
layers as an empty entry. This was causing users to have weird effects,
such as setting a Phi Limit from 0-90 defaulting back to -90,90 or tick
boxes reticking themselves.

A helper was added for the model layers to unify this behaviour of what
is an acceptable falsey type.

The presenter layer was updated to only filter out empty strings and
none types, so that Qt doesn't choke on them.

The view was updated to remove logic (which shouldn't be in the view)
that was doing a final check before setting the value.