Make bounding box calculation optional
For some instruments we do not have the full geometry set up so using
the bounding box is not possible. In this case it is ok to assume the
pixel edges lie mid-way between the pixel centres. However for
instruments where the bounding box is known, it could be beneficial to
use the bounding box, to avoid projecting counts into regions where
there should be none.

Re #23012