Change the reference angle for the projection
When summing in Q, there was previously a workaround for the reference
angle used in the projection in order to make it work with ConvertUnits for
the final conversion to Q. Instead, use our own conversion to Q (using RefRoi so
we can specify the correct angle). Then we can simply use the reference
detector angle for the projection. We just need to ensure the same angle is used in the
conversion to Q to match.

Re #23012