Remove workaround from unit test for reversed Q workspace
The lambda range fix fixes a weird problem in the unit tests where we
were getting a negative value for the first lambda bin edge. This meant
that ConvertUnits was not recognising that the workpace in Q needed
reversing so the output in Q bins were the wrong way round. They are now
the correct way round so we can remove the workaround in the unit test
for the incorrect direction.

Re #23012