Delay sliceviewer refresh from ADS update
Updating the model when an MDEventWorkspace is changed
requires running BinMD. The ADS observer
is fired as part of the algorithm execution and for
replacement the algotithm as not yet released the
workspace lock so trying to run another algorithm
throws an exception. We avoid this by delaying the
view update until the next event loop cycle and the
observer has then completed.