Update the system test and results with new behaviour
In theory, I would expect to be able to produce the same results as
previously, so this is not ideal. But to do that would require moving
the instrument very precisely to the correct position and this would
probably require a calibration file (which we don't currently have an
easy way of applying in Mantid). This is beyond the scope of this PR
because this is a very old system test which will probably be superseded
in the future anyway.

For now, I've therefore made an approximate change
to the instrument position, which makes the output look roughly
sensible, but the counts are different, so I have had to update the
reference files as well. This should be fine for now for the purposes of
testing just to ensure we catch any behaviour changes, even though this
example is unfortunately not such a good example scientifically any more.

Re #20327