Create a new version of CtRQ
Add a version number to the underlying classes to switch between the old
and new behaviours

Replace the original test and data and move the new functionality to a
v2 test. Note that there was already a v2 of the data so the new
reference files are v3 even though they relate to v2 of the algorithm.

Note that the new calculation is tested in the ReflectometryTransform
tests so does not need to be tested in v2 of ConvertToReflectometryQ2.

Re #20327