Temporary fix for WorkspaceSelector bug
This commit is a workaround for a bug where the WorkspaceSelector's blank row (which
is required to make the selection optional) is removed when the ADS is
cleared. This causes a problem in the reflectometry GUI because the
workspace selection for the flood workspace becomes mandatory, and when a workspace is next
created it becomes the flood workspace, causing the reduction to fail.

In future this will be fixed by ensuring the blank row always exists in
the WorkspaceSelector. However, for release 6.0.0 I do not want to
change that widget late in the release cycle. Hence this
workaround is being added instead, which refreshes the WorkspaceSelector
on the reflectometry GUI after the ADS has been cleared. This ensures that the blank row gets
re-added and solves the reflectometry bug without risking effects in
other parts of the code.

This commit will be reverted when the correct fix is added.

Re #30317