Abins: recalculate if setting was changed
Make a small extension to the existing cache validation system so that
data will be recalculated if the instrument setting was changed.

This isn't ideal because sometimes all that needs to be re-done is
broadening. But a bunch of work is needed to defer broadening, so this
will have to happen later / separately from this introduction of the
settings system.