Update status of fit in fit browser during Seq Fit
Provides user with a visual prompt to indicate a fit has been performed. This meant exposing a public signal of Fitbrowser to python. Have also updated test, including checking notifier passes right argument post seq. fit - which required making a deepcopy of the fit output (this was not necessary for the actual usage of the UI, it is only for test purposes).