Remove 2D instruments
Some of the supporting infrastructure is left behind, but generally
the purpose here is to strip back the changes a bit. This branch still
covers iteration over angles, Instrument settings, Lagrange and
pushing a bunch of the Abins algorithm into a parent class. Quite a
lot as-is, would be good to review/merge while solving problems with 2D.