Abins: add output/expt unit selection
Sometimes experimental data is provided in meV. There are also a
couple of "gotchas" for end-users in converting results with Mantid's
ConvertUnits dialogue: EMode must be set to Indirect, incident energy
is not relevant, and "DeltaE" is the choice for meV while
"DeltaE_inWavenumber" is the choice for cm-1.

As a convenience, provide an Abins option in familiar units. The
selected units will also be assumed for any experimental data file
provided to the same Abins call.