Abins: Fix existing cache logic for setting changes
The current logic is _intended_ to ensure that a change in instrument
setting triggers a recalculation of structure factors. This wasn't
working correctly, and instead the data was _always_ recalculated
because the bugged comparison would always fail.

This was difficult to spot for small tests because the notice about
recalculating data was hidden at "information" level; this is promoted
to the "notice" logging level.

There is still room for improvement; Lagrange monochromator changes
_shouldn't_ trigger a recalculation because the _unbroadened_
structure factors are still correct. But TOSCA forward/backward
detector changes should trigger a recalculation because they change
the energy/q relationship. A more sophisticated scheme is needed.