Add tests for new unit covnersion and fix fit table tests
Fit table tests now different because the peak centres converted to dSpacing are now matrix workspaces (and some changes to how these are stroed in the fit results of the model).
Tests around unit conversion are subject to change in near future - presently only difc can be calculated from the workspace.