Abins: add instrument names to Settings
Requested by Sanghamitra Mukhopadhyay

e.g. 'Cu(111)' becomes 'Cu(111) (Lagrange)'

This is rather verbose and a bit redundant, but gives a better user
experience from the GUI in Workbench.

Ideally the user interface would adapt and only offer settings for the
selected instrument; this goes beyond the capabilities of an
automatically-generated Algorithm GUI and would require an Interface
to be developed.