Resolve some unhelpful behaviour of parameters in SetSample
When SetSample was run from the Algorithm dialog in the workbench UI
it wasn't possible to leave some of the dictionary parameters blank.
This is a valid use case for some combinations of parameters

Also when clearing a parameter that previously had a value the
"blanking out" didn't take effect straight away. You had to reopen
the algorithm dialog and try to run the algorithm a second time to
see validation errors relating to empty fields

Fix problem with instrument name in unit test where test failed
if you happened to have your preferred facility set to TEST_LIVE