Add log value check before workspace name comparison
This fixes a crash on RHEL7 at the cell function because
the first column of the result fit workspace did not contain
workspace names, but log values instead. This caused issues
when casting the data to a string for comparison. This adds
a check to make sure the column is SourceName before trying
to find the appropriate workspace.