SimulatedDensityOfStates fix for .phonon files with LOTO splitting
The current PHONON_HEADER_REGEX uses the wrong column for 'weight'
when there is LO-TO splitting and the 'q-pt=' line contains seven
floats. There is a seven-float example given with the regex, but in
this case the weight happens to be the same number as the z-coordinate
of the splitting vector.

The weight should be the *fourth* float value in all cases, which is
not always the last one. The regex groups and float capture have been
adjusted accordingly.