Fix seg fault in MonteCarloAbsorption
MonteCarloAbsorption would occasionally crash when running on Ubuntu
when a material on one of the shapes wasn't defined. In this situation
the code in one of the threads would create the missing material. Sometimes
multiple threads would create the missing material and it appears this
wasn't thread safe.
I have move the creation of the dummy material into the constructor of
CSGObject instead of it happening on demand in CSGObject::material