Fix bug in ass calculation in PaalmanPingsMonteCarloAbsorption
The container shape was previously still present in the ass calculation.
This would have affected the attenuation even if no scatter points were
generated in the container.
Fix this by taking a copy of the sample and container shape at the start
of the algorithm and then swap them in as required
Doing this required some methods on the API::Sample class to be exposed
to Python. This included one (setEnvironment) that previously took a
unique_ptr as a parameter. I don't think you can expose a method that has
a unique_ptr parameter using boost python so I've changed this to be a
shared_ptr. Any performance hit from using shared_ptr rather than unique_ptr
will be better than copying the environment object which is what would
otherwise be required