Support loading event transmission runs in SANS
Loading transmission runs was gated off as it required "additional
work". This is because the raw files load in monitors using
LoadNexusMonitors, which will not work on processed files.

Instead we now select the correct nexus period, from the output
multiperiod file, then pack it into the expected workspaces. This is
despite the fact that it's workspace monitors but stil...

This does not support multi period files, as some additional thought is
required on how to get the "correct" period number, and to be frank, I
don't have the data to hand. So let's cross that bridge later and
instead throw an error that we can't do it