SANS Don't notify for inserted blank tables
Don't notify whenever a blank row is inserted. Instead we should only
notify from the table model if the user has changed the contents.

Trying to notify prevents us (in the next commit) from inserting an
empty row as it will try to update the view to say "Hey I've got a row"
despite the fact the view is already adding a row as it's the only place
it can come from.

This should also slightly help performance of us not having to reload
file info each time