Notify when a row is entered before trying to use it
Adds a step to notify the presenter we have inserted a row, before then
trying to update the contents. This prevents a race where the table
model gets asked about a row it doesn't know about.

This should prevent a crash where hitting enter then load will cause an
IndexError