Fix SANS cached Can details not using wav range for key
Fixes the SANS Cached can details not using the current wavelength range
for the key. This means that every cache entry is a miss, and can cause
random workspaces to be selected as a candidate can instead.

This was primarily seen when running a reduction with RangeLin
wavelength slicing twice and the save set to Both, as a random WS would
be selected causing the save alg to choke.

Unfortunately, we can't add an automatic test as the API currently
doesn't have a way of setting wavlength slices unlike the GUI. When it
was originally written there was no advantage of doing your wavelength
slices in one single call vs in an outer loop. This feature request has
also caught us out on the user-base side where weirdly the GUI is
significantly faster than the API, because the API exposes less.

When we go to add an API we'll need to add a test that runs the system
test twice to confirm it's fixed. However, as we'll already be writing a
system test that runs it once (to verify the API) this will be trivial
to do.