Re #31298 Muon Results Table Now Remembers Previous Fits
The problem was when clearing the cache the list of fits performed were also being removed. I have removed this from the clear cache function. Additionally added clear fits when undo fits is pressed and added a subsciber so the results table is cleared when undo fits is pressed, as you can't output any results once undo fits is pressed