Re #30526 Fix Fit Line Being Same Color As Data Line
The fit line color would match the raw data line color if you plotted more lines than the number of colors. This was because the color queue would reset once the extra lines were added. When you swap to the fitting tab all curves are removed except one. As lines are removed their colors are added back to the color queue which would result in the fit line being the same color as the data line since every color was added back.