Fix LARMOR trying to scale LAB (and HAB) fields
Fix LARMOR trying to scale LAB fields into mm. This was changed in 5.10
but was only picked up recently. One of the fields is used as an angle
input, so cannot be sensibly scalled by 1000.

In this case we will leave the tolerance...etc. as mm as they are
unaffected and instead tweak the LAB (and HAB) values specifically