Set IntensityScaling when initialising CrystalFieldMultiSite
When combining CrystalField objects into a CrystalFieldMultiSite the IntensityScaling parameters have to be set explicitly.
In addition, when a scaling factor is applied to any but the first CrystalField objects it was ignored. This is also fixed now.

The IntensityScaling parameters from the first CrystalField object are used to set the spectrum IntensityScaling parameters of the CrystalFieldMultiSite object.
The scaling factors for the CrystalField objects are used as the ion IntensityScaling parameters of the CrystalFieldMultiSite object.