Replace on_update with update_centre in test
on_update rows should not been invoked when a user file is updated, as
it is not a row.
Instead copy in the centre values then ask the presenter for the beam
centre finder to updates its view.

This test was missed during the earlier commit, and for the reviewers
sanity will go in its own commit