Fix save can not working on SANS
Fixes Save Can option not working since the wavelength loops
refactoring. This was missing a nested object so would throw an
attribute error.

Unfortunately, it appears there was no automated test either. I've added
one but long term this test class needs to be split. This would allow us
to run multiple tests in parallel, or in this case, just the one.
However, as were approaching release this needs to be left to the future