Added warning for IntensityScaling value mismatch
When combining two CrystalField objects with IntensityScaling settings only the IntensityScaling values of the first object are passed on to the CrystalFieldMultiSite object, the IntensityScaling values of the second CrystalField object are ignored.
In case there is a difference between the IntensityScaling settings of the two CrystalField objects a warning is output in order to make the user aware of this.