Add QThread/Python threading adaptor
Adds a widget that adapts between Python threading and QThreads. This
provides a number of advantages described in the docstring for
associated classes.

The short-version is it gives us (much) better stack traces, allows us
to get rid of some bare catch exceptions and avoids users having to wire
up Qt signals and slots to simply run something off the GUI thread.

Whilst it's not present in this commit, this has been tested with the
ISIS SANS interface to verify the functionality with Qt and also fix up
some API usability problems noticed whilst implementing this