Add setting for unit test mode to AsyncQtAdaptor
This will manually change the object under test to run synchronously,
rather than having to faff with trying to get threading right and
possibly having race conditions.

We could include some special logic to manually fire the callbacks on
completion, as per the Qt adaptor layer. However, this encourages people
to write big tests. Instead we encourage them in the docs to separate
the testing for the async method and the testing for the callback(s)