Remove warning from instrument viewer
When viewing instruments that have banks that are width "one detector"
in x or y or z, the instrument viewer was giving a warning about "failed
to build unwrapped surface" when you viewed in one of the cylindrical
projections. Remove the warning and also the code to assign values to
u_min and u_max shortly after. The pre-existing values of 0 and 0 are
fine for these values which are in radians
Also include some line length changes that come from the latest clang
format version