Remove proton charge integration when impossible
Event data processing algorithms assume in multiple places that event
data workspaces should have an associated proton charge, when it is not
the case for ILL data.

In those places, whether there is actually a proton charge to integrate
is check before trying to do so, so as to silence multiple unwarranted
warnings.