Moved interruption point to after sleep
Now a thread checks for the cancel signal invoked by another
thread after its sleep time, instead of checking before sleep as it was
until now. This change is not expected to affect adversly anything.
This change is done in the hope that MonitorLiveDataTest failures of os-x
build will disappear or atleast significantly go down relative to the current
frequency of failures