Overwrite logs when loading if needed
When loading an event nexus with the LoadAllLogs option checked, some
logs may be loaded twice, thus sending a warning.
It is better to overwrite by default in the general algorithm, as in
most cases there shouldn't be anything to ovrewrite, and in this case,
the version coming from the general algorithm is likely to be better
than the one copied by LoadAllLogs.