Shim interface rendering surface for testing
Adds in a shim so that we work with an interface in the Instrument view
instead of a concrete type.

This is the lowest impact way of prepping for tests, if the heirarchy is
wrong it will fail to build of open at all. It also means we can start
building tests without touching the implementation from this point
onwards.

One unfortunate problem is the fact the interfaces inherit from Qt
types, this is the least impact way to get this working through.