Fix up tests with failing GMock assertions
Fixes various tests that were failing silently. For most it was simply a
case or change a call.

However the AlgorithmProgress* tests requires a significant rewrite. It
is undefined behaviour interacting with a mock then setting
expectations. By setting the presenter to be a member of the mock this
interleaving broken said mocks.

This required a rewrite to completely seperate out the presenter and
mock injection, despite the complex interdependencies.