Fix race-y tests
Fixes a race-y test that won't report if two updates come in with < 0.1s
between them.
Using a sleep is, well poor practice. But without trying to inject some
timer mechanism this more closely emulates the fact a Qt Slot may fire
after some delay.

Additionally, I've removed testing with a real algorithm, as we lose
control of the Signal Slot mechanism and it can lead to unreliable
tests. Instead since we are testing the slot directly if this fails we
know it's the signal or Qt (more likely the signal).