Check if outputWS is event after flip re #31820
During a workspace multiply, if we have SingleValue*Event then the operation can be done in
place, meaning the output gets cast to the LHS. BinaryOperation knows the multiplication is
commutative, so it flips the operation to Event*SingleValue. If we cast
the output to the LHS before the flip, the output is a SingleValue but
the LHS becomes and Event, which caused errors.