Fix check if path variable is empty
CMake fills variables of filepath type with a
string that contains -NOTFOUND to indicate
the path has not been found by a search.
Comparing this as a string yields a positive
because there is a string it's just a special
string to indicate a missing value.
Checking the variable directly forces CMake
to check for the special marker string and give
the correct answer