Properly handle wildcards in functions that get their data from the...
description of the function, for example np.asarray get's the call tip from this description. If the wildcard handling is not present dtype appears required and the wildcard is optional where in reality it is the reverse.

Also refactor out generate_call_tip as a separate function to reduce complexity for Flake8 compatibility.