Use 6 nearest neighbours for WISH
Change the value of the NumberOfNeighbours parameter to 6 instead
of the default value of 8. The centres of the spectra on WISH are
arranged in such a way that the 8th and 9th nearest neighbour can
be v close together meaning the top 8 neighbours are sometimes
different on different platforms due to tiny differences in the floating
point results. This change aligns the behaviour on all platforms