Abins: cache a_traces
The a_trace data is smaller than the eigenvectors, so the memory
footprint of caching them is not too bad. Accessing them in several
places with this get() method is a bit cleaner than passing around the
processed data.

No cache invalidation logic is included; no set() methods are
provided, which hopefully makes it clear that PowderData is not meant
to be changed once instantiated.