Fix issue with checking for grid on variable in matplotlib dict
For a set of 3d axes there is no variable gridOn in the dictionary, meaning we get a python error. Unfortunately there doesn't seem to be a nice way of determining whether the grid on, without accessing some private data of the axes class. The fix for now is to specialise the calls for mantid axes (projection=mantid and projection=mantid3d).