Fix iviewer bug on replacing ws
While attempting to fix this bug, the previous commit introduced an
issue, and didn't even fix it properly.

The correct way is to wait for previous signals to finish their work
before attempting to compute the new miniplot. This way it has data to
work on, and no need to try to get this data before it is ready.